Fixed problem with files outside of working dir
Previous patch was based upon the (false) assumption that
inotify_event->name is the file path, however it just contains the
filename. This made it only work for files within the current working
directory. In order to know where a particular inotify event came from we
need to use the watch descriptor (wd) that we were given when we
initally called add_watch.

Also removed unnecessary check that inotify event is a IN_MODIFY event,
since we only recieve IN_MODIFY events anyway.